+++ heapRestaurant.cpp
@@ -4,7 +4,7 @@ heapRestaurant::heapRestaurant()
 {
 }
 
-heapRestaurant::heapRestaurant(restaurant *a, long long b)
+heapRestaurant::heapRestaurant(restaurant *a, double b)
 {
     item = a;
     distance = b;
